+++ tests/ui/no_comma_after_clone.rs
@@ -1,0 +1,10 @@
+use dst_factory::make_dst_factory;
+
+#[make_dst_factory(create, clone no_std)]
+struct NoCommaAfterClone {
+    id: i32,
+    data: str,
+}
+
+fn main() {
+}